chore(clippy): adopt clippy 1.95 lint suggestions
Toolchain bump to rust-clippy 1.95.0 surfaced 8 dormant lint violations
across 6 files; mechanical fixes applied via `cargo clippy --fix` plus
3 manual `sort_by_key` rewrites that clippy flagged as non-machine-
applicable.

- librarian/preview/memory.rs, logging.rs, tools/file_group.rs:
  `sort_by(|a,b| b.x.cmp(&a.x))` -> `sort_by_key(|b| Reverse(b.x))`
  (unnecessary_sort_by)
- retrieval/sync.rs, tools/symbol/call_graph/traversal.rs:
  drop redundant `.into_iter()` on `IntoIterator`-accepting args
  (useless_conversion)
- tools/markdown/edit_markdown.rs: drop `usize as usize` no-op casts
  (unnecessary_cast)

Tracker entry F-6 in docs/trackers/bug-fix-session-log.md records the
toolchain-bump observation and the load-bearing test-fixture follow-up.

+++ src/tools/markdown/edit_markdown.rs
@@ -48,7 +48,7 @@ pub fn perform_section_edit_ext(
         resolve_section_range(content, heading_query).map_err(|e| anyhow::anyhow!("{}", e))?;
 
     let lines: Vec<&str> = content.split('\n').collect();
-    let heading_idx = (range.heading_line - 1) as usize;
+    let heading_idx = range.heading_line - 1;
     let end_idx = compute_section_end(&lines, heading_idx + 1, range.level);
 
     match action {
@@ -225,7 +225,7 @@ pub fn find_consumed_subsections(content: &str, heading_query: &str) -> Result<V
         resolve_section_range(content, heading_query).map_err(|e| anyhow::anyhow!("{}", e))?;
 
     let lines: Vec<&str> = content.split('\n').collect();
-    let heading_idx = (range.heading_line - 1) as usize;
+    let heading_idx = range.heading_line - 1;
     let end_idx = compute_section_end(&lines, heading_idx + 1, range.level);
 
     let mut in_code_block = false;
@@ -320,7 +320,7 @@ pub(crate) fn perform_scoped_edit(
         resolve_section_range(content, heading_query).map_err(|e| anyhow::anyhow!("{}", e))?;
 
     let lines: Vec<&str> = content.split('\n').collect();
-    let heading_idx = (range.heading_line - 1) as usize;
+    let heading_idx = range.heading_line - 1;
     let end_idx = compute_section_end(&lines, heading_idx + 1, range.level);
 
     // Extract the section content (heading + body) with trailing newline
